requirements.txt
django
django_registration
django-termsandconditions
django-markdown2
django-mdmail
django-middleware-global-request
python-frontmatter
markdown2
django-ninja
